refactor : Validator 코드 통합
